Changed sum type to float64

+++ models/categorySummary.go
@@ -5,5 +5,5 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type CategorySummary struct {
 	ID           primitive.ObjectID `json:"categoryId,omitempty" bson:"_id,omitempty"`
 	CategoryName string             `json:"categoryName,omitempty"`
-	Sum          float32            `json:"sum,omitempty"`
+	Sum          float64            `json:"sum,omitempty"`
 }
